backend/requirements.txt
fastapi
uvicorn
crewai
langchain-openai
python-dotenv
pydantic
